Use cairo_translate to set cairo output coordinates to match global coordinates
This simplifies rendering, since we don't have to subtract the origin of
the output from the selection coordinates for each operation.

This also fixes the coordinates for the crosshairs drawn on the screen.
Previously, the calls to draw the crosshairs used a mix of both
coordinate systems, which would work for a single output, but not for
many multi-output setups.